refactor board content structure
