+++ src/pages/Boards/BoardContent/ListColumns/Column/ListCards/Card/Card.jsx
@@ -1,0 +1,58 @@
+import { Card as MuiCard } from '@mui/material'
+import CardActions from '@mui/material/CardActions'
+import CardContent from '@mui/material/CardContent'
+import CardMedia from '@mui/material/CardMedia'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+
+import GroupIcon from '@mui/icons-material/Group'
+import CommentIcon from '@mui/icons-material/Comment'
+import AttachmentIcon from '@mui/icons-material/Attachment'
+function Card({ temporaryHideMedia }) {
+  if (temporaryHideMedia) {
+    return (
+      <MuiCard
+        sx={{
+          cursor: 'pointer',
+          boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
+          overflow: 'unset'
+        }}
+      >
+        <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
+          <Typography>Alpaca</Typography>
+        </CardContent>
+      </MuiCard>
+    )
+  }
+  return (
+    <MuiCard
+      sx={{
+        cursor: 'pointer',
+        boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
+        overflow: 'unset'
+      }}
+    >
+      <CardMedia
+        sx={{ height: 140 }}
+        image="https://scontent.fhan19-1.fna.fbcdn.net/v/t39.30808-6/439586236_3664752570520746_2486317449040107640_n.jpg?_nc_cat=109&ccb=1-7&_nc_sid=6ee11a&_nc_eui2=AeH6zeg7s2TmZLG98iHiembXRSS7wPxRTNBFJLvA_FFM0M2BKWhJZu2oDq-GEUAnOG5mzQR_NwvqWqKR3-kIHEmO&_nc_ohc=j4jb4Vh7g7AQ7kNvgEnaUjk&_nc_ht=scontent.fhan19-1.fna&_nc_gid=ANZWWzFRWodjNbEC1IweY8m&oh=00_AYBqO_9ZQVL1yeUM1tdUGEF6Cbq-Yz8aWFlgLLvqDxwGhw&oe=66E96542"
+        title="green iguana"
+      />
+      <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
+        <Typography>Alpaca</Typography>
+      </CardContent>
+      <CardActions sx={{ p: '0 4px 8px 4px' }}>
+        <Button size="small" startIcon={<GroupIcon />}>
+          20
+        </Button>
+        <Button size="small" startIcon={<CommentIcon />}>
+          15
+        </Button>
+        <Button size="small" startIcon={<AttachmentIcon />}>
+          10
+        </Button>
+      </CardActions>
+    </MuiCard>
+  )
+}
+
+export default Card
