+++ src/pages/Boards/_id.jsx
@@ -1,18 +1,14 @@
 // board details
 
 import Container from '@mui/material/Container'
-import AppBar from '~/components/AppBar'
-import BoardBar from './BoardBar'
-import BoardContent from './BoardContent'
+import AppBar from '~/components/AppBar/AppBar'
+import BoardBar from './BoardBar/BoardBar'
+import BoardContent from './BoardContent/BoardContent'
 
 function Board() {
   return (
     <>
-      <Container
-        disableGutters
-        maxWidth={false}
-        sx={{ height: '100vh' }}
-      >
+      <Container disableGutters maxWidth={false} sx={{ height: '100vh' }}>
         <AppBar />
         <BoardBar />
         <BoardContent />
